fix: Fix commons version 3.5.0 and fix longValue in shouldGetTransactionTotalAmountV2 test

+++ src/test/java/it/pagopa/transactions/utils/TransactionsUtilsTest.java
@@ -226,7 +226,7 @@ class TransactionsUtilsTest {
                         it.pagopa.ecommerce.commons.documents.PaymentNotice::getAmount
                 ).sum();
         Long methodTotalAmount = utils.getTransactionTotalAmount(transaction);
-        assertEquals(totalAmount, methodTotalAmount.intValue());
+        assertEquals(totalAmount, methodTotalAmount.longValue());
     }
 
     @Test
